requirements.txt
PyGObject>=3.42.0
pycairo
dnspython
